Migrate the CPU stress test to the chaosapi (#24)
* Migrate the CPU stress test to the chaosapi

This commit migrates the stress_cpu script completely to the chaosapi.
This commit serves as a first discussion of the new feature. Let's
discuss in the PR whether changes are necessary and relevant. If
everything is okay the next scripts will be moved:

* Are the function docs appropriate and should there be any information
that states that a working internet connection is necessary to make use
of the scripts?
* Is the design and architecture okay

* Removed number of cores